Improve Windows-friendliness for MSVC builds.
The only essential dependencies are GNU make for Windows and bash
with some tools (diff, grep...). bash and GNU tools are provided
by git for Windows distributions.

This commit doesn't overwrite the LIB environment variable (which
is crucial for the MS linker) and assumes cl.exe to be in a directory
in the PATH environment variable rather than a hardcoded path to
VS 2010.

A binary distribution of GNU make for Windows (from 2006) is available
at http://gnuwin32.sourceforge.net/packages/make.htm - you'll need the
binaries and dependencies ZIPs.
Some scripts require GNU diff to support `--strip-trailing-cr`, which
wasn't the case for the ancient diff.exe (v2.7) provided by my git
distribution. A newer version (2.8.7) can be downloaded from
http://gnuwin32.sourceforge.net/packages/diffutils.htm.